requirements.txt
aerich==0.7.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
dictdiffer==0.9.0
fastapi==0.78.0
flake8==6.0.0
flake8-commas==2.1.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
iso8601==1.1.0
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.10.0
pydantic==1.10.11
pyflakes==3.0.1
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
starlette==0.19.1
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.7.1
uvicorn==0.17.6
watchgod==0.8.2
websockets==11.0.3
yarl==1.9.2
